+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ## [Unreleased]
 
+## [6.2.0] - 2025-12-20
+
 ### Added
 - Avalonia: Added file-dialogs to `IWindowManager`
 - WinUI: Added file-dialogs to `IWindowManager`
@@ -336,7 +338,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 ### Added
 - Initial release of Caliburn.Light where at least WPF can be considered stable
 
-[Unreleased]: https://github.com/tibel/Caliburn.Light/compare/v6.1.0...HEAD
+[Unreleased]: https://github.com/tibel/Caliburn.Light/compare/v6.2.0...HEAD
+[6.2.0]: https://github.com/tibel/Caliburn.Light/compare/v6.1.0...v6.2.0
 [6.1.0]: https://github.com/tibel/Caliburn.Light/compare/v6.0.0...v6.1.0
 [6.0.0]: https://github.com/tibel/Caliburn.Light/compare/v5.3.0...v6.0.0
 [5.3.0]: https://github.com/tibel/Caliburn.Light/compare/v5.2.1...v5.3.0